modify size_t,warmachine/ add calloc,strdup,substr

+++ libft/ft_bzero.c
@@ -6,13 +6,15 @@
 /*   By: ylee <ylee@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/08 15:44:02 by ylee              #+#    #+#             */
-/*   Updated: 2020/10/08 15:47:45 by ylee             ###   ########.fr       */
+/*   Updated: 2020/10/12 13:44:28 by ylee             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	ft_bzero(void *s, int n)
+#include <stddef.h>
+
+void	ft_bzero(void *s, size_t n)
 {
-	int		i;
+	size_t	i;
 	char	*str;
 
 	i = 0;
